# PERF: Don't copy if already sorted in DataSplitter