- [ ] closes #51077  (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

```
       before           after         ratio
     [2a983bb1]       [2a980997]
     <datasplitter-ordered-nocopy~1>       <datasplitter-ordered-nocopy>
+      14.1±0.4ms       32.8±0.3ms     2.32  groupby.ApplyDictReturn.time_groupby_apply_dict_return
-     3.56±0.03ms      3.01±0.07ms     0.85  groupby.Apply.time_sorted_scalar_function_single_col(5)
-            236M             197M     0.83  groupby.Apply.peakmem_sorted_scalar_function_multi_col(6)
-            220M             182M     0.83  groupby.Apply.peakmem_sorted_scalar_function_single_col(6)
-        51.4±4ms       34.1±0.5ms     0.66  groupby.Apply.time_sorted_scalar_function_single_col(6)
```
Benchmarks.
The regression in apply_dict_return is real, and I think the ``.copy()`` method has a lot of overhead(both the finalize + name validation I think).
No memory improvement is showing up there since the size isn't big enough.