@jbrockmendel 2023-03-16
would #51109 address some or all of the finalize/name slowdown you mentioned?