@lithomas1 2023-03-31
> Closes #43011?

For the non CoW case, yes. This is already optimized for CoW (since a no-op take there will return a shallow copy).